+++ Services/ApiKeyManagement/Discord/ApiKeyCommand.cs
@@ -1,14 +1,26 @@
-﻿using DSharpPlus;
-using DSharpPlus.Entities;
-using DSharpPlus.SlashCommands;
-using System.Drawing;
-using System.Globalization;
-using System.Runtime.CompilerServices;
-using TornBot.Services.Players.Service;
-using TornBot.Services.TornApi.Services;
-using TornBot.Services.TornStatsApi.Services;
+﻿//
+// TornBot
+//
+// Copyright (C) 2024 TornBot.com
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+// 
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
 
-namespace TornBot.Services.Discord.Commands.Slash
+using DSharpPlus.SlashCommands;
+using TornBot.Services.Players.Service;
+
+namespace TornBot.Services.ApiKeyManagement.Discord
 {
     public class ApiKeyCommand : ApplicationCommandModule
     {
@@ -39,14 +51,11 @@ namespace TornBot.Services.Discord.Commands.Slash
             [Option("apikey", "Add torn api key with public access level")] string api_key)
             //[Option("TornStatsApiKey", "Add torn stats api key - not required (you can input 0)")] string TornStats_api_key)
         {
-            
             api_key = api_key.Trim();
             string api_key_respons = "";
             
-                api_key_respons = _players.AddApiKey(api_key);
+            api_key_respons = _players.AddApiKey(api_key);
             
-            
-
             await ctx.CreateResponseAsync(api_key_respons, ephemeral: true);
