+++ Services/Database/DatabaseContext.cs
@@ -1,12 +1,28 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿//
+// TornBot
+//
+// Copyright (C) 2024 TornBot.com
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+// 
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using System.Reflection;
-using TornBot.Database.Entities;
-using static System.Net.Mime.MediaTypeNames;
-using MySql.Data.MySqlClient;
+using TornBot.Services.Database.Entities;
 
-namespace TornBot.Database
+namespace TornBot.Services.Database
 {
     public class DatabaseContext : DbContext
     {
